--- packages/eas-cli/src/submissions/android/types.ts
@@ -1,11 +1,0 @@
-import { SubmissionContext, SubmitCommandFlags } from '../types';
-
-export interface AndroidSubmitCommandFlags extends SubmitCommandFlags {
-  type?: 'apk' | 'aab';
-  key?: string;
-  androidPackage?: string;
-  track: string;
-  releaseStatus: string;
-}
-
-export type AndroidSubmissionContext = SubmissionContext<AndroidSubmitCommandFlags>;
